+++ docs/changes.rst
@@ -7,6 +7,8 @@ Version 0.4.0
 To be released.
 
 - Geofront becomes to require Paramiko 2.0.0 or higher.
+- :mod:`geofront.util` is gone now.  Geofront instead becomes to require
+  :mod:`typing` and :mod:`tsukkomi`.
 
 
 Version 0.3.2